+++ content-archive.php
@@ -1,5 +1,10 @@
 <?php
 	if(have_posts()) {
+    	?>
+    	<div class="row">
+        	<div class="flex-wrap">
+            	
+    	<?php
 		while(have_posts()) {
 			the_post();
 			
@@ -11,12 +16,34 @@
 			} else {
 				$edit_link = "";
 			}
-				
+
+			if(is_home()):
+                $thumb_id = get_post_thumbnail_id();
+                $thumb_url_array = wp_get_attachment_image_src($thumb_id, 'blog-featured-image', true);
+                $thumb_url = $thumb_url_array[0];
+
+                if(empty($thumb_url) || $thumb_url == SITEURL . '/wp-includes/images/media/default.png')  {
+                    $thumb_url = 'http://placehold.it/737x737/';
+                }
+
+			endif;
 			
 			echo "<article>";
+			if(!empty($thumb_url)):
+			?>
+			<a href="<?=$permalink?>">
+			<img src="<?=$thumb_url?>" alt="<?=get_the_title()?>" class="blog-pic" />
+			</a>
+			
+			<?php
+			endif;
 			the_title("<h3><a href=\"$permalink\">", "</a>$edit_link</h3>");
-			the_excerpt();
+			sherpa_excerpt();
 			echo "</article>";
 		}
+		?>
 		
+        	</div>
+    	</div>
+		<?php
 	}