+++ inc/theme-features.php
@@ -18,6 +18,8 @@ if ( ! function_exists('sherpa_theme_features') ) {
 		add_theme_support( 'title-tag' );
 		
         add_image_size( 'testimonial-profile-pic', 450, 450, TRUE );
+        add_image_size( 'blog-featured-image', 737, 414, TRUE );
+        add_image_size( 'blog-interior-image', 1170, 200, TRUE );
 	}
 	
 	// Hook into the 'after_setup_theme' action