+++ page-home.php
@@ -4,16 +4,6 @@
 	*/
 	get_header(); ?>
 
-<section id="home-slideshow">
-	<div class="container">
-		<div class="row">
-			<div class="<?=FULLWIDTH?>">
-				<img src="http://placehold.it/1140x385/0640aa/FFFFFF/&text=Slider+to+go+here" />
-			</div>
-		</div>
-	</div>
-</section>
-
 <section id="home-main">
 	<div class="container">
 		<div class="row">